Move base grapher to more generic position
