+++ app/friend_graphs/psycopg_grapher.py
@@ -7,7 +7,7 @@ from app import APP_ENV
 from app.models import DATABASE_URL, USER_FRIENDS_TABLE_NAME
 from app.decorators.datetime_decorators import logstamp
 from app.decorators.number_decorators import fmt_n
-from app.friend_graphs.base_grapher import BaseGrapher, DRY_RUN, BATCH_SIZE, USERS_LIMIT
+from app.base_grapher import BaseGrapher, DRY_RUN, BATCH_SIZE, USERS_LIMIT
 
 class PsycopgGrapher(BaseGrapher):
     def __init__(self, dry_run=DRY_RUN, batch_size=BATCH_SIZE, users_limit=USERS_LIMIT,
